Refactor pfm-http
- separate things
- move config and context into global static
