+++ pfm-http/src/routes/forex_handlers/mod.rs
@@ -1,0 +1,3 @@
+pub(super) mod convert;
+pub(super) mod rates;
+pub(super) mod timeseries;
