+++ pfm-http/src/routes/root_handlers/healthcheck.rs
@@ -1,0 +1,5 @@
+use axum::{http::StatusCode, response::IntoResponse};
+
+pub(crate) async fn ping_handler() -> impl IntoResponse {
+    (StatusCode::OK, "OK")
+}
