+++ pfm-http/src/routes/root_handlers/mod.rs
@@ -1,0 +1,1 @@
+pub(super) mod healthcheck;
